requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
psycopg2-binary==2.9.11
redis==5.0.8
bcrypt==4.2.0
PyJWT==2.9.0
google-generativeai==0.8.3
python-docx==1.1.2
